refactor(utils.projen-builder): extract `BuildStep` to `build-steps` module
