+++ packages/utils/projen-builder/src/index.ts
@@ -1,24 +1,10 @@
 import { withDefaults } from '@arroyodev-llc/utils.projen'
-import { type Project, type ProjectOptions } from 'projen'
+import { type BuildStep } from './build-step.ts'
 import type {
 	GConstructor,
 	GenericProjectConstructor,
 	ProjectConstructorOptions,
-	TypedPropertyDescriptorMap,
 } from './types.ts'
-
-export abstract class BuildStep {
-	declare abstract outputOptionsType: {}
-	declare abstract outputType: {}
-
-	abstract applyOptions(
-		options: ProjectOptions
-	): ProjectOptions & this['outputOptionsType']
-
-	abstract applyProject(
-		project: Project
-	): TypedPropertyDescriptorMap<this['outputType']>
-}
 
 export class ProjectBuilder<
 	T extends GenericProjectConstructor = GenericProjectConstructor,
